refactor(core): promote shard types and chart derivations out of the app
format, diff, pareto, neighbours, arms, metrics and requests only ever
depended on @atlas/core types, but lived in app/src/util — so any second
frontend would have had to duplicate them and drift. They now live in core
next to the types they operate on, together with the shard contract
(Manifest/Registry/IndexRow/...) and the fixtureRow test builder. The app
keeps its import paths through a shim in data/types.ts; everything else
imports @atlas/core directly.
