+++ app/src/components/cell-drawer.ts
@@ -5,8 +5,8 @@ import type { HeatCell, PossibleCell } from '../data/derive.js';
 import type { IndexRow } from '../data/types.js';
 import { href } from '../router.js';
 import { store } from '../store.js';
-import { fmtInt } from '../util/format.js';
-import { headlineMetric } from '../util/metrics.js';
+import { fmtInt } from '@atlas/core';
+import { headlineMetric } from '@atlas/core';
 import { addButton } from './add-modal.js';
 import { AtlasElement } from './base.js';
 import { icon } from './icons.js';
