+++ app/src/data/fixture.ts
@@ -1,5 +1,6 @@
+export { fixtureRow } from '@atlas/core';
 /** A tiny registry/index fixture for tests. */
-import type { CompiledIndexRow, SiteConfig } from '@atlas/core';
+import type { SiteConfig } from '@atlas/core';
 import siteJson from '../../../site/config.json';
 import type { Registry } from './types.js';
 
@@ -144,20 +145,3 @@ export function fixtureRegistry(): Registry {
   };
 }
 
-export function fixtureRow(over: Partial<CompiledIndexRow> = {}): CompiledIndexRow {
-  return {
-    run_id: 'cfg0000000000001--serve-single-i256-o256-v1--abc123',
-    cell_id: 'cell',
-    config_id: 'cfg0000000000001',
-    workload_id: 'serve-single-i256-o256-v1',
-    kind: 'serving',
-    engine: { id: 'vllm', version: '0.27.1', minor: '0.27' },
-    model: { id: 'Qwen/Qwen3-8B', quant_id: 'fp8' },
-    hardware: { id: 'nvidia-rtx-4090', count: 1 },
-    metrics: { output_tok_s: 120.5, ttft_p50: 80, ttft_p95: 120 },
-    provenance: { login: 'someone', submitted_at: '2026-08-23T12:00:00Z' },
-    verification_level: 'self-reported',
-    path: 'results/vllm/Qwen/Qwen3-8B/nvidia-rtx-4090/x.json',
-    ...over,
-  };
-}
