+++ app/src/views/run-view.ts
@@ -43,13 +43,13 @@ import {
 import type { IndexRow } from '../data/types.js';
 import { href, modelHref, navigate } from '../router.js';
 import { store } from '../store.js';
-import { armDiff, armLabel, cellArms, prefillPoints } from '../util/arms.js';
+import { armDiff, armLabel, cellArms, prefillPoints } from '@atlas/core';
 import { cssVar, seriesColor } from '../util/colors.js';
 import { absDateTime } from '../util/dates.js';
-import { metricDelta } from '../util/diff.js';
-import { fmtGB, fmtInt, fmtMs, fmtNum, fmtPct, fmtTokS, fmtW, shortSha } from '../util/format.js';
-import { blockCards, headlineMetric } from '../util/metrics.js';
-import { requestSamples } from '../util/requests.js';
+import { metricDelta } from '@atlas/core';
+import { fmtGB, fmtInt, fmtMs, fmtNum, fmtPct, fmtTokS, fmtW, shortSha } from '@atlas/core';
+import { blockCards, headlineMetric } from '@atlas/core';
+import { requestSamples } from '@atlas/core';
 import { fmtDefault, isDefault } from '../components/param-form.js';
 import { modelRefFor } from './explore-view.js';
 import { ViewElement } from './view-base.js';
