+++ packages/core/src/format.ts
