+++ packages/core/src/index.ts
@@ -44,3 +44,13 @@ export { computeCoverage, emptyCell, isReleaseVersion, minorsBehind } from './co
 export type { CoverageRegistry, CoverageOptions } from './coverage.js';
 export { buildPacket, renderFlags, renderServeCommand, AGENT_RULES } from './packet.js';
 export type { PacketSpec, PacketRegistry, PacketEngineEntry, PacketModelEntry } from './packet.js';
+
+export * from './shards.js';
+export * from './format.js';
+export * from './diff.js';
+export * from './pareto.js';
+export * from './neighbours.js';
+export * from './arms.js';
+export * from './metrics.js';
+export * from './requests.js';
+export { fixtureRow } from './fixtures.js';
