+++ packages/core/src/pareto.test.ts
