+++ packages/core/src/pareto.ts
